+++ index.d.ts
@@ -6,6 +6,7 @@
 */
 
 export function header(source: string): string[] | undefined;
+export function body(source: string): string;
 export function parse(source: string, options?: Partial<ParserSettings>): DocumentFragment;
 // The `bind` function yields the elements below.
 //   1. Added or removed child elements of the target container element (<any> excluding the others).