+++ src/parser/api.ts
@@ -1,4 +1,5 @@
 export { header } from './api/header';
+export { body } from './api/body';
 export { parse } from './api/parse';
 export { bind } from './api/bind';
 export { caches } from './api/cache';