+++ index.d.ts
@@ -18,6 +18,7 @@ export function render(target: HTMLElement, options?: RenderingOptions): void;
 export function toc(source: DocumentFragment | HTMLElement | ShadowRoot): HTMLUListElement;
 export function info(source: DocumentFragment | HTMLElement | ShadowRoot): Info;
 export function context(base: DocumentFragment | HTMLElement | ShadowRoot, bound?: string): (el: Element) => boolean;
+export function normalize(source: string): string;
 
 export interface ParserSettings {
   // For servers.