+++ src/parser/api.ts
@@ -2,4 +2,5 @@ export { parse } from './api/parse';
 export { bind } from './api/bind';
 export { header } from './api/header';
 export { body } from './api/body';
+export { normalize } from './api/normalize';
 export { caches } from './api/cache';