+++ src/parser/api/parse.ts
@@ -4,7 +4,7 @@ import { eval } from '../../combinator';
 import { segment } from '../segment';
 import { header } from '../header';
 import { block } from '../block';
-import { normalize } from '../normalize';
+import { normalize } from './normalize';
 import { figure } from '../../util/figure';
 import { footnote } from '../../util/footnote';
 import { frag } from 'typed-dom';